#59 Moves Model*TagLib and *Renderer classes and their dependencies out of the .merit package, as they are not merit-specific any more
