+++ grails-app/taglib/au/org/ala/biocollect/ModelTagLib.groovy
@@ -1,4 +1,4 @@
-package au.org.ala.biocollect.merit
+package au.org.ala.biocollect
 
 import grails.converters.JSON
